Added bylaws and minor footer updates
Added Bylaws as PDF for download. We are trying to be open so we
have published our bylaws so anyone can read why has SPy o.z. been
founded and also what are our internal rules.

Minor updates in footer, disabled target _blank in link to self :)
and removed https from link to chat as it is https only so user
will be redirected anyway and without https in link looks better in
mobile.